Split metatensor's C++ header into multiple files
