+++ scripts/check-c-api-docs.py
@@ -8,9 +8,12 @@ import sys
 from pycparser import c_ast, parse_file
 
 
-ROOT = os.path.join(os.path.dirname(__file__), "..")
+ROOT = os.path.realpath(os.path.join(os.path.dirname(__file__), ".."))
 C_API_DOCS = os.path.join(ROOT, "docs", "src", "core", "reference", "c")
-FAKE_INCLUDES = os.path.join(ROOT, "python", "scripts", "include")
+FAKE_INCLUDES = [
+    os.path.join(ROOT, "python", "scripts", "include"),
+    os.path.join(ROOT, "scripts", "include"),
+]
 METATENSOR_HEADER = os.path.relpath(
     os.path.join(ROOT, "metatensor-core", "include", "metatensor.h")
 )
@@ -61,7 +64,9 @@ def functions_in_outline():
 
 
 def all_functions():
-    cpp_args = ["-E", "-I", FAKE_INCLUDES]
+    cpp_args = ["-E"]
+    for path in FAKE_INCLUDES:
+        cpp_args += ["-I", path]
     ast = parse_file(METATENSOR_HEADER, use_cpp=True, cpp_path="gcc", cpp_args=cpp_args)
 
     functions = []
